Switch packet constructors to take a ByteArraySegment instead of a byte[]. Fixes an issue with ethernet packets that have trailing bytes that a user reported on the packet.net sourceforge project site.
The previous implementation assumed that a packet was present at a given offset into
a byte[]. The implicit assumption was that all of the bytes after the given
offset and until the end of the byte[] were a part of that packet or packets
nested inside of that packet. This isn't the case however. Some ethernet packets
for instance can have trailing bytes which implies that their size may be smaller
than the remaining bytes in the byte[].

Modify the IPv4 and IPv6 packets to adjust the payload size based on the ip header
information. This allows the tcp, udp etc packets to be smaller than the remaining
bytes in the packet byte[].

Implemented a UdpTest unit test for the case of an ethernet packet with
trailing bytes. The udp checksum and length are correct now that the
ByteArraySegment ByteLength values are adjusted based on the ip headers.

Removed the byte[] constructors for all packets. Users can construct packets via
var blah = new SomePacket(new ByteArraySegment(theUsersByteArray)); It was getting
difficult to ensure that all packet constructors had a form that took byte[] and
removing the byte[] constructors removes a large number of lines of code.

Modified the Packet class to automatically set its PosixTimeval based on the
current time and removed the PosixTimeval parameter from all of the packet constructors.
The timeval is often only used when capturing packets or reading them from an
on-disk file so requiring the timeval field to be set for each constructed packet
was likely an unnecessary requirement for end users of the library.