Remove Packet.Timeval and all mention or use of Timeval. Also remove PosixTimeval and the PosixTimeval unit test.
Evan Plaice has pointed out that packets don't really have an intrinsic timeval. The timeval
property was a carryover from the act of capturing the packet and is unrelated to the role
of Packet.Net as purely a packet parser. Users that desire associating a timeval with a packet
can easily encapsulate the two in a container class.