Changes to support Raw IP packets as written by Wireshark and Rawcap.exe(http://www.netresec.com/?page=RawCap)
Raw IP packets are IPv4 and IPv6 directly inside the raw libpcap packet, the first
byte of the Raw IP packet is the first byte of the IPv4 or IPv6 packet.

Search for 'LINKTYPE_RAW' on: http://www.tcpdump.org/linktypes.html

This is the file format you get when you use Rawcap.exe to capture packets on the loopback interface.
More: http://www.netresec.com/?page=RawCap

Winpcap cannot normally capture off of the loopback interface, so this is a nice tool to capture loopback packets, and with this support
for Raw linklayer these can now be parsed.

It is quite possible this is the link layer used by Wireshark and other tools on Linux when capturing
from the loopback interface, although they can use LinuxSLL as an option.