FrameControlField.cs - Create a new property, 'Protected', and flag 'Wep' as obsolete to better reflect the current use of this flag
Originally the flag was 'Wep' because this was the encryption standard in use at the time.
Since then there are a number of different encryption schemes that are used and this
flag is generally known as 'protected' or 'data protected' instead.

Create a new 'Protected' properly and flag the 'Wep' one as obsolete to encourage moving
away from it so we can remove 'Wep' in the future.

Update unit tests to use the 'Protected' properly instead of 'Wep'.