Reduce Lazy overhead
LazyThreadSafetyMode.PublicationOnly is set because this is enough for the purpose here, as it's not very likely multiple threads will access the payload at the same time on initial hit (even if it would, the underlying ByteArraySegment would remain the same across threads). Compared to LazyThreadSafetyMode.ExecutionAndPublication (the default), this saves about 25% time on create + initial hit. On gbit speeds this is noticable (I saw the Lazy initialization come up in profiler).